# TimePeriod::overlaps performance improvements