--- a/benchmark/Aeon/Calendar/Benchmark/Gregorian/TimePeriodBench.php
+++ b/benchmark/Aeon/Calendar/Benchmark/Gregorian/TimePeriodBench.php
@@ -14,17 +14,23 @@ use Aeon\Calendar\Gregorian\TimePeriod;
  */
 final class TimePeriodBench
 {
+    private TimePeriod $firstPeriod;
+
+    private TimePeriod $secondPeriod;
+
+    public function __construct()
+    {
+        $this->firstPeriod = new TimePeriod(DateTime::fromString('2020-01-01 00:00:00'), DateTime::fromString('2020-01-06 00:00:00'));
+        $this->secondPeriod =  new TimePeriod(DateTime::fromString('2020-01-04 00:00:00'), DateTime::fromString('2020-01-08 00:00:00'));
+    }
+
     public function bench_distance() : void
     {
-        (new TimePeriod(DateTime::fromString('2020-01-01 00:00:00'), DateTime::fromString('2020-01-06 00:00:00')))
-            ->distance();
+        $this->firstPeriod->distance();
     }
 
     public function bench_overlaps() : void
     {
-        (new TimePeriod(DateTime::fromString('2020-01-01 00:00:00'), DateTime::fromString('2020-01-06 00:00:00')))
-            ->overlaps(
-                new TimePeriod(DateTime::fromString('2020-01-04 00:00:00'), DateTime::fromString('2020-01-08 00:00:00'))
-            );
+        $this->firstPeriod->overlaps($this->secondPeriod);
     }
 }