### Before
```
22 subjects, 110 iterations, 110 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.400 [27,863.516 27,911.023] 0.400 (μs)
⅀T: 3,064,986.800μs μSD/r 115.578μs μRSD/r: 2.180%
suite: 1343d13bd3d3c5d171c5a256030ee14cbb5def63, date: 2020-07-23, stime: 08:37:10
+---------------------+-----------------------------------------------------------+------+-----+-----------+
| benchmark           | subject                                                   | revs | its | mean      |
+---------------------+-----------------------------------------------------------+------+-----+-----------+
| InitializationBench | bench_datetime_immutable                                  | 5    | 5   | 0.015ms   |
| InitializationBench | bench_aeon_year                                           | 5    | 5   | 0.061ms   |
| InitializationBench | bench_aeon_month                                          | 5    | 5   | 0.110ms   |
| InitializationBench | bench_aeon_day                                            | 5    | 5   | 0.169ms   |
| InitializationBench | bench_aeon_timezone                                       | 5    | 5   | 0.369ms   |
| InitializationBench | bench_aeon_time                                           | 5    | 5   | 0.038ms   |
| InitializationBench | bench_aeon_datetime_from_datetime_immutable               | 5    | 5   | 0.884ms   |
| InitializationBench | bench_aeon_datetime_create                                | 5    | 5   | 0.679ms   |
| InitializationBench | bench_aeon_datetime_new                                   | 5    | 5   | 0.672ms   |
| InitializationBench | bench_aeon_datetime_from_string                           | 5    | 5   | 0.863ms   |
| DayIterationBench   | bench_iteration_over_last_half_of_the_year_multiple_times | 5    | 5   | 608.973ms |
| DayBench            | bench_time_between                                        | 5    | 5   | 0.063ms   |
| DayBench            | bench_to_datetime_immutable                               | 5    | 5   | 0.002ms   |
| DayBench            | bench_is_equal                                            | 5    | 5   | 0.000ms   |
| DayBench            | bench_is_before                                           | 5    | 5   | 0.004ms   |
| DayBench            | bench_is_after                                            | 5    | 5   | 0.004ms   |
| DayBench            | bench_is_before_eq                                        | 5    | 5   | 0.004ms   |
| DayBench            | bench_is_after_eq                                         | 5    | 5   | 0.004ms   |
| DateTimeBench       | bench_unix_timestamp                                      | 5    | 5   | 0.003ms   |
| DateTimeBench       | bench_to_datetime_immutable                               | 5    | 5   | 0.003ms   |
| TimePeriodBench     | bench_distance                                            | 5    | 5   | 0.016ms   |
| TimePeriodBench     | bench_overlaps                                            | 5    | 5   | 0.062ms   |
+---------------------+-----------------------------------------------------------+------+-----+-----------+
```

### After

```
22 subjects, 110 iterations, 110 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 0.400 [21,142.375 21,184.905] 0.400 (μs)
⅀T: 2,325,661.200μs μSD/r 91.303μs μRSD/r: 1.332%
suite: 1343d13d5bc66218e83ed43cd9333d445ae6ffe5, date: 2020-07-23, stime: 08:38:03
+---------------------+-----------------------------------------------------------+------+-----+-----------+
| benchmark           | subject                                                   | revs | its | mean      |
+---------------------+-----------------------------------------------------------+------+-----+-----------+
| InitializationBench | bench_datetime_immutable                                  | 5    | 5   | 0.015ms   |
| InitializationBench | bench_aeon_year                                           | 5    | 5   | 0.059ms   |
| InitializationBench | bench_aeon_month                                          | 5    | 5   | 0.112ms   |
| InitializationBench | bench_aeon_day                                            | 5    | 5   | 0.171ms   |
| InitializationBench | bench_aeon_timezone                                       | 5    | 5   | 0.374ms   |
| InitializationBench | bench_aeon_time                                           | 5    | 5   | 0.038ms   |
| InitializationBench | bench_aeon_datetime_from_datetime_immutable               | 5    | 5   | 0.899ms   |
| InitializationBench | bench_aeon_datetime_create                                | 5    | 5   | 0.694ms   |
| InitializationBench | bench_aeon_datetime_new                                   | 5    | 5   | 0.696ms   |
| InitializationBench | bench_aeon_datetime_from_string                           | 5    | 5   | 0.871ms   |
| DayIterationBench   | bench_iteration_over_last_half_of_the_year_multiple_times | 5    | 5   | 461.086ms |
| DayBench            | bench_time_between                                        | 5    | 5   | 0.064ms   |
| DayBench            | bench_to_datetime_immutable                               | 5    | 5   | 0.002ms   |
| DayBench            | bench_is_equal                                            | 5    | 5   | 0.000ms   |
| DayBench            | bench_is_before                                           | 5    | 5   | 0.002ms   |
| DayBench            | bench_is_after                                            | 5    | 5   | 0.002ms   |
| DayBench            | bench_is_before_eq                                        | 5    | 5   | 0.002ms   |
| DayBench            | bench_is_after_eq                                         | 5    | 5   | 0.002ms   |
| DateTimeBench       | bench_unix_timestamp                                      | 5    | 5   | 0.001ms   |
| DateTimeBench       | bench_to_datetime_immutable                               | 5    | 5   | 0.003ms   |
| TimePeriodBench     | bench_distance                                            | 5    | 5   | 0.011ms   |
| TimePeriodBench     | bench_overlaps                                            | 5    | 5   | 0.030ms   |
+---------------------+-----------------------------------------------------------+------+-----+-----------+
```